+++ test/debug_test.dart
@@ -1,12 +1,20 @@
 import 'package:dartbag/debug.dart';
+import 'package:path/path.dart' as pathlib;
 import 'package:test/test.dart';
 
 void main() {
   test(
-    'currentDartPackagePath',
+    'currentDartFilePath:',
     onPlatform: {'browser': const Skip()},
     () {
-      expect(currentDartPackagePath(), 'test/debug_test.dart');
+      var path = currentDartFilePath();
+      expect(pathlib.isAbsolute(path), true);
+      expect(path, contains(pathlib.context.separator));
+      expect(normalizeSeparators(path), endsWith('/test/debug_test.dart'));
+
+      path = currentDartFilePath(packageRelative: true);
+      expect(path, contains(pathlib.context.separator));
+      expect(normalizeSeparators(path), 'test/debug_test.dart');
     },
   );
 
@@ -25,3 +33,6 @@ void main() {
     expect(assertsEnabled, true);
   });
 }
+
+/// Normalizes directory separators to the POSIX style.
+String normalizeSeparators(String path) => path.replaceAll(r'\', '/');